dvd_data_checker/requirements.txt
# DVD Data Checker Dependencies
# Core database connectivity
psycopg==3.2.1

# Configuration management
PyYAML==6.0.1

# Data analysis and manipulation
pandas==2.1.4
numpy==1.26.2

# Date and time handling
python-dateutil==2.8.2

# Logging and monitoring
colorlog==6.7.0

# HTTP requests for webhooks (optional)
requests==2.31.0

# CSV handling
csvkit==1.0.7

# JSON handling
jsonschema==4.19.1

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development dependencies
black==23.7.0
flake8==6.0.0
mypy==1.5.1
isort==5.12.0

# Documentation
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0
